Consolidate github tool
