+++ src/marvin/tools/github.py
@@ -1,9 +1,49 @@
+from datetime import datetime
+from typing import List, Optional
+
 import httpx
+from pydantic import BaseModel, Field, validator
 
 import marvin
 from marvin.tools import Tool
-from marvin.utilities.models.github import GitHubIssue
 from marvin.utilities.strings import slice_tokens
+
+
+class GitHubUser(BaseModel):
+    """GitHub user."""
+
+    login: str
+
+
+class GitHubComment(BaseModel):
+    """GitHub comment."""
+
+    body: str = Field(default="")
+    user: GitHubUser = Field(default_factory=GitHubUser)
+
+
+class GitHubLabel(BaseModel):
+    """GitHub label."""
+
+    name: str = Field(default="")
+
+
+class GitHubIssue(BaseModel):
+    """GitHub issue."""
+
+    created_at: datetime = Field(...)
+    html_url: str = Field(...)
+    number: int = Field(...)
+    title: str = Field(default="")
+    body: Optional[str] = Field(default="")
+    labels: List[GitHubLabel] = Field(default_factory=GitHubLabel)
+    user: GitHubUser = Field(default_factory=GitHubUser)
+
+    @validator("body", always=True)
+    def validate_body(cls, v):
+        if not v:
+            return ""
+        return v
 
 
 class SearchGitHubIssues(Tool):
